Incorrect placement of plugin.
I believe this is incorrectly used here. This will cause various issues, especially using it with Android Studio. This is also the reason you have to specify the QA compile error. The plugin `com.android.support:exifinterface:+` should at the very least be used but then not even, rather have it as part of the installation process to include the dependency build.gradle.